requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
decorator==4.4.2
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
moviepy==1.0.3
numpy==2.1.2
pillow==11.0.0
proglog==0.1.10
PySide6==6.8.0.1
PySide6_Addons==6.8.0.1
PySide6_Essentials==6.8.0.1
requests==2.32.3
setuptools==75.1.0
shiboken6==6.8.0.1
tqdm==4.66.5
urllib3==2.2.3
wheel==0.44.0
